@@ -4,7 +4,7 @@ version: 1.7.0
 homepage: https://jaweii.github.io/Flutter_beautiful_popup/example/build/web
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.18.2 <3.0.0'
 
 dependencies:
   flutter:
