@@ -71,7 +71,23 @@ class TemplateThumb extends BeautifulPopupTemplate {
         ),
       );
       final minHeight = 40.0 - (outline ? 2 : 0);
-      return RaisedButton(
+      return ElevatedButton(
+          onPressed: onPressed,
+          child: Container(
+            constraints: BoxConstraints(
+              minWidth: 100,
+              minHeight: minHeight,
+            ),
+            alignment: Alignment.center,
+            child: Text(
+              label,
+              style: TextStyle(
+                color: Colors.white.withOpacity(0.95),
+              ).merge(labelStyle),
+            ),
+          ));
+
+      /*RaisedButton(
         color: Colors.transparent,
         elevation: elevation,
         highlightElevation: 0,
@@ -97,7 +113,7 @@ class TemplateThumb extends BeautifulPopupTemplate {
           borderRadius: BorderRadius.circular(50),
         ),
         onPressed: onPressed,
-      );
+      );*/
     };
   }
 
